+++ src/time_split_app/new/my_extensions.py
@@ -1,4 +1,7 @@
-"""Dummy extensions for the Time Split application."""
+"""Dummy extensions for the Time Split application.
+
+See https://time-split.readthedocs.io/en/latest/api/time_split.app.reexport.html for documentation.
+"""
 
 from typing import Any
 
@@ -9,7 +12,7 @@ from time_split.app import create_explorer_link
 from time_split.types import DatetimeIndexSplitterKwargs
 
 from time_split_app.widgets import DataLoaderWidget
-from time_split_app.widgets.parameters import SplitterKwargsWidget
+from time_split_app.widgets.parameters import ExpandLimitsWidget, ScheduleWidget, SpanWidget, SplitterKwargsWidget
 
 
 class MyDatasetLoader(DataLoaderWidget):
@@ -60,6 +63,7 @@ class MyDatasetLoader(DataLoaderWidget):
 
         index = pd.date_range(start, end)
         data = {"dummy-data": [1] * len(index)}
+
         return pd.DataFrame(data, index=index)
 
 
@@ -72,7 +76,12 @@ def my_plot_fn(*args: Any, **kwargs: Any) -> Axes:
 
 def my_select_fn() -> DatetimeIndexSplitterKwargs:
     """Returns keyword arguments for https://time-split.readthedocs.io/en/stable/api/time_split.html#time_split.split."""
-    return SplitterKwargsWidget().select_params()
+    return SplitterKwargsWidget(
+        schedule_widget=ScheduleWidget(),
+        before_widget=SpanWidget(span="before"),
+        after_widget=SpanWidget(span="after"),
+        limits_widget=ExpandLimitsWidget(),
+    ).select_params()
 
 
 def my_link_fn(*args: Any, **kwargs: Any) -> str: